feat: Enhance WeChat bill import functionality and improve database configuration
- Updated database connection settings to support utf8mb4 charset.
- Modified settings to include charset in database URL.
- Refactored budget API import to a separate module.
- Changed ImportLog model to use BigInteger for user_id.
- Implemented WeChat bill parsing for both CSV and XLSX formats.
- Enhanced column mapping detection for WeChat bill data.
- Improved error handling and user feedback during file import.
- Updated frontend components to support XLSX file uploads.
- Added new API endpoints for WeChat import operations.
- Improved UI for account and transaction management.
- Removed obsolete environment configuration file.

+++ backend/app/models/import_log.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, DateTime, Enum, JSON, ForeignKey, Text
+from sqlalchemy import Column, Integer, BigInteger, String, DateTime, Enum, JSON, ForeignKey, Text
 from sqlalchemy.sql import func
 from sqlalchemy.orm import relationship
 from app.config.database import Base
@@ -15,7 +15,7 @@ class ImportLog(Base):
     __tablename__ = "import_logs"
 
     id = Column(Integer, primary_key=True, autoincrement=True, comment="导入日志ID")
-    user_id = Column(Integer, ForeignKey("users.id", ondelete="CASCADE"), nullable=False, comment="用户ID")
+    user_id = Column(BigInteger, ForeignKey("users.id", ondelete="CASCADE"), nullable=False, comment="用户ID")
     source = Column(String(50), nullable=False, comment="数据来源(wechat/alipay/manual)")
     file_name = Column(String(200), comment="导入文件名")
     file_size = Column(Integer, comment="文件大小(字节)")
